Single Match Results for Demo
This uses Alan's demo script to give results on the single match page.
You need to type in a number and postocde present on the AWS test ES,
e.g 7 EX2 6GA - the search will fire on completion and the expanded
address returned. Clicking the Search button will either give you the
"no address" error if the box is empty, or timeout  as the API is not
there yet.