Override universalBin task for server to include parser resources
This change allows the removal of all the 'shell' steps needed after
construction of the zip package via SBT. These manual steps include:
    1. Modifying the zip using 'jar uf' to add parser resources (details
below)
    2. Creating a copy of the zip to allow using a predicatable name for
the Jenkins Cloud Foundry plugin to push the app bundle.

Removing the need for the shell steps will make this build more portable
and easily reproducible.

Note that this required upgrading sbt-native-packager from v.1.1.4 to
v1.3.3

Before the server (address-index-api/server) can be deployed to Cloud
Foundry, a zip package is created using:
 $ sbt "project address-index-server" "universal:packageBin"

and the zip modified by using a command like so:
    $ jar uf server/target/universal/address-index-server.zip \
     -C parsers/src/main/resources libcrftagger-linux.so \
     -C parsers/src/main/resources addressCRFA.crfsuite \
     -C parsers/src/main/resources/input_pre_post_processing business \
     -C parsers/src/main/resources/input_pre_post_processing company \
     -C parsers/src/main/resources/input_pre_post_processing county \
     .....

Ideally, it is desired that all steps required to create a distribution
be done entirely by the build tool, otherwise these steps would have to
be replicated every time a package has to be made (either by devs or by
Jenkins).