Fix default encoding being used for Jenkins causing a test failure
The default encoding being picked up by the JVM on jen-m Jenkins was
causing an error while reading a file containing an accented character.
The default encoding for Travis CI builds (and for local execution of
tests) was UTF-8, however for the Jenkins JVM it was not.

Instead of a code change to specify the encoding which might have impact
elsewhere, I chose to specify the encoding only for the specific SBT
task ('test') which obviates the need to evaluate the impact elswhere.

Task: REG-1985